feat(#173): extract Queue route from legacy App.tsx into separate component (Phase 2)
- Extract QueueTab component from 17k+ line legacy App.tsx into frontend/src/pages/Queue/index.tsx
- Move all helper functions (formatDuration, sortRows, normalizeSortValue, sortStateNext, SortTh, Collapse, Stat) to the new module
- Create frontend/src/pages/Queue/__tests__/QueueTab.test.tsx with React Testing Library tests
- Update frontend/src/legacy/App.tsx to import QueueTab from new location instead of defining inline
- Removes 1145 lines from legacy App.tsx while maintaining full functionality
- No breaking changes: Queue tab behavior identical before/after extraction
- Validates Phase 2 extraction process for remaining routes in #173

This is a complete, self-contained route extraction as specified in the 4-phase migration plan.
